fix: initialized connect API for edit course page

+++ server/controllers/adminController.js
@@ -200,7 +200,7 @@ export const getCourse = async (req, res) => {
       sub_lesson_videos: courseData.rows[0].video_directory,
     }
   );
-  console.log(filesMetaData);
+  // console.log(filesMetaData);
 
   return res.json({
     data: courseData.rows[0],
@@ -219,9 +219,8 @@ export const getCourse = async (req, res) => {
 export const updateCourse = async (req, res) => {
   const courseId = req.params.courseId;
   const adminId = req.query.adminId;
-  const action = req.body.action;
   const mediaFiles = req.files;
-  console.log(mediaFiles);
+  console.log(req.files.course_attached_files);
 
   const updatedCourse = {
     courseName: req.body.course_name,
@@ -286,13 +285,41 @@ export const updateCourse = async (req, res) => {
       message: "updated successfully 1",
     });
   } else {
-    const filesPublicIdForDelete = [];
-    const jsonsOfAllFiles = JSON.parse(req.body.all_media);
-    for (let i of jsonsOfAllFiles) {
-      filesPublicIdForDelete.push(JSON.parse(i).public_id);
+    const courseMediaFiles = await pool.query(
+      `
+    SELECT courses.cover_image_directory, courses.video_trailer_directory, sub_lessons.video_directory
+    FROM courses
+    INNER join lessons
+    ON lessons.course_id = courses.course_id
+    INNER JOIN sub_lessons
+    ON sub_lessons.lesson_id = lessons.lesson_id
+    WHERE courses.course_id = $1 AND admin_id = $2
+    `,
+      [courseId, adminId]
+    );
+
+    let filesPublicIdForDelete = [];
+
+    // If there are attached files
+
+    if (req.files.course_attached_files != undefined) {
+      const courseAttachedFiles = await pool.query(
+        `
+      SELECT * from files where course_id = $1
+      `,
+        [courseId]
+      );
+
+      for (let file of courseAttachedFiles.rows) {
+        filesPublicIdForDelete.push(JSON.parse(file.directory).public_id);
+      }
     }
 
-    //delete everything from cloudinary
+    for (let publicId of Object.values(courseMediaFiles.rows[0])) {
+      // console.log(JSON.parse(publicId).public_id);
+      filesPublicIdForDelete.push(JSON.parse(publicId).public_id);
+    }
+
     for (let filePublicId of filesPublicIdForDelete) {
       await cloudinaryUpload(filePublicId, "delete");
     }
@@ -310,33 +337,33 @@ export const updateCourse = async (req, res) => {
       "course_video_trailers"
     );
 
-    for (let file of req.files.course_attached_files) {
-      updatedCourse.courseAttachFiles.push(
-        JSON.stringify(
-          await cloudinaryUpload(file, "upload", "course_attached_files")
-        )
-      );
-    }
-
-    let fileName = [];
-    let fileType = [];
-    let fileSize = [];
-
-    // console.log(req.files.course_attached_files);
-
-    for (let file of req.files.course_attached_files) {
-      fileName.push(file.originalname);
-      fileType.push(file.mimetype);
-      fileSize.push(file.size);
-    }
-
-    // console.log(updatedCourse.courseAttachFiles);
-
     updatedSubLesson.subLessonVideo = await cloudinaryUpload(
       ...req.files.sub_lesson_videos,
       "upload",
       "sub_lesson_videos"
     );
+
+    let fileName = [];
+      let fileType = [];
+      let fileSize = [];
+
+    if (req.files.course_attached_files != undefined) {
+    
+
+      for (let file of req.files.course_attached_files) {
+        updatedCourse.courseAttachFiles.push(
+          JSON.stringify(
+            await cloudinaryUpload(file, "upload", "course_attached_files")
+          )
+        );
+
+        console.log(updateCourse.courseAttachFiles);
+
+        fileName.push(file.originalname);
+        fileType.push(file.mimetype);
+        fileSize.push(file.size);
+      }
+    }
 
     // update the files' directory except for attached files
     await pool.query(
@@ -382,20 +409,29 @@ lesson_update as (
       ]
     );
 
-    await pool.query(
-      `
-    DELETE from files WHERE course_id = $1
-    `,
-      [courseId]
-    );
 
-    await pool.query(
-      `
-    INSERT INTO files (course_id, file_name, type, size, directory)
-    VALUES ( $1, unnest($2::text[]), unnest($3::text[]), unnest($4::int[]), unnest($5::text[]));
-    `,
-      [courseId, fileName, fileType, fileSize, updatedCourse.courseAttachFiles]
-    );
+    if (req.files.course_attached_files != undefined) {
+      await pool.query(
+        `
+      DELETE from files WHERE course_id = $1
+      `,
+        [courseId]
+      );
+
+      await pool.query(
+        `
+      INSERT INTO files (course_id, file_name, type, size, directory)
+      VALUES ( $1, unnest($2::text[]), unnest($3::text[]), unnest($4::int[]), unnest($5::text[]));
+      `,
+        [
+          courseId,
+          fileName,
+          fileType,
+          fileSize,
+          updatedCourse.courseAttachFiles,
+        ]
+      );
+    }
 
     return res.json({
       message: "updated successfully",
